Add CommandParser class + corresponding test class
We need a CommandParser class to interpret the command and run the
right methods after that.

Having a class with all these methods together allow for more
organized code and easier debugging, as all the code relevant to
this functionality is in the same class.

Let's create a CommandParser class that consolidates all the
methods required for interpreting the incoming commands and running
the appropriate methods from other classes.