Troubleshoot Gradle check
The Gradle check is not passing.

The Javadoc comment are dangling, as the methods have been
commented out. This is because the other classes required in the
CommandParser class have not been implemented yet.

Let's change the Javadoc comment to a standard comment temporarily,
and check if Gradle accepts it.