Implement user input interpretation
We need a way to take the user input and convert it into
information that the parser can understand.

These methods will help to isolate the command portion of the user
input and allow for easier calling of the appropriate commands
later.

Let's,
* create a few methods to split up the user input into the
command portion and the information portion.
* EXTRA: remove irrelevant files that were mistakenly committed
and merged.