Implement command calling from other classes
The methods in the other system classes cannot be called from the
main BusynessManager class.

We need to connect the BusynessManager class to the methods in
other system classes for BusynessManager to perform its tasks and
functions.

Let's create some methods in the parser to take in the command
from the user input, decide on what the command is, and run the
appropriate method from the other system classes.