Modify command extraction method + Edit JUnit tests
The command extraction does not work for single-word user inputs.
Moreover, the JUnit tests are not as organized as desired.

Fixing the command extraction will expand the range of usability to
encompass single-word user inputs that do not contain any information.
Organising the JUnit tests will help to clarify the different
scenarios that the methods can face.

Let's,
* Fix extractCommand() to get single-word user inputs working.
* Organise the JUnit tests for better clarity.