Rename custom exceptions and modify InvalidStringException
The custom exceptions are not named as exceptions. Moreover, the
message printed in InvalidString is repeated in all scenarios.

Adding the word "Exception" to the custom exceptions will
explicitly inform others that they are exceptions. Removing the
message parameter will help to simplify the code.

Let's,
* Rename all custom exceptions and their class files to include
the word "Exception"
* Modify InvalidString such that there is no more message
parameter in the constructor.