Integrate other required classes into parser code
The code utilising the other classes are commented out, as these
classes had not been implemented yet.

Integrating them into the code will allow these classes to be
called, thereby enabling the parser to execute the methods present
in these classes.

Let's uncomment the rest of the code that utilises the other
required classes.