Add assertions for some parser methods
For some of the command calling methods, the possibility of
negative values for quantity and price can cause the code to
receive incorrect inputs, resulting in the code not working as
expected.

Assertions can help in this aspect by preventing the code from
receiving negative values for quantity and price. Thus, our code
is kept safe from these potentially destructive values.

Let's,
* add assertions in the command calling methods that require
quantity and price to be parameters.
* EXTRA: uncomment printProducts(). (because I forgot).