Add logging to parser class
The parser class does not contain any logging of what is going on
in the class.

Logging will help during debugging to find potential issues with
the code. Logging can also help to explain the flow of the code
within the parser class.

Let's add some logging ability to the parser class.