Add enumeration to fix invalid businessType bug
There is a bug where businessType is not limited to RETAIL or FNB.
The code allows any String to work as a businessType.

This may lead to unpredictable behaviour in the code, as the code
does not expect the businessType to be anything else except FNB and
RETAIL.

Let's,
* Create a enumeration for businessType, so that only RETAIL and
FNB are allowed.
* Modify the businessType input code to combat against incorrect
inputs.
* Convert any remaining Strings created to constants.