Clean up code for v1.0 release
The code is at the final stage before release, and it is about time
to release a .jar file for BusynessManager.

Releasing the .jar file gives our product to users to use. Checking
the code before that will be a good habit which helps to make sure
that our code is good before releasing.

Let's
* Comb through the code and check for any irregularities
* Edit the build.gradle file to generate the .jar file
* Update any UI or JUnit tests if necessary