Add incorrect input handling for variable extraction
When an incorrect input is given, some extraction methods do not
give an indication that the extraction has failed to the user.
Moreover, all the extraction methods do not repeat the input
instruction message to the user after an incorrect input.

Giving an error message will tell the user that their input is
wrong, and they need to try again. Repeating the instruction
message helps to remind users of what they need to input.

Let's,
* Add checks with error messages for the scenario when there is no
input for name and password extraction.
* Add the ability for the instruction message to loop whenever the
extraction is unsuccessful.
* Modify the error message that is given when there is no input
from the user.