Add confirmation check for first-time setup
When the business ID is invalid, the first-time setup instantly
starts without checking with the user if he/she wants to start it.

Checking with the user first allows the user to cancel the first-
time setup and terminate the program. This is in the scenario where
he/she may have accidentally put in an incorrect business ID.

Let's
* Implement a confirmation check to start the first-time setup.