Modify automated text UI tests
Since the sequence of instructions for initialising BusynessManager
is now different, the input and expected text files for the
automated text UI tests are outdated.

Updating these test files will help to check if the updated
functionality works as expected.

Let's update the automated text UI test files to include the new
functionality.