Fix invalid product ID bug
When an invalid product ID is inputted for some methods, the
program will crash with a NullPointerException. This is because the
invalid product ID does not exist as a key in the productList hash
map.

Adding proper handling of this error will allow the user to
continue using the program, even when the invalid product ID is
inputted.

Let's,
* Handle this invalid product ID bug.
* Add new constants for the error handling.
* Modify the text UI tests for regression testing.