Fix JUnit tests
The Gradle check is failing, as some of the JUnit tests' actual
outputs do not match the expected outputs.

Fixing these JUnit tests will help to check if the methods work as
expected, assuring us that the code is not buggy.

Let's modify the JUnit tests' expected outputs, and see if the
Gradle check passes.