Create product ID detection
To input the product ID, the user needs to type the correct and
entire format of the ID. (i.e. ID_XXXX)

Simplifying this step to only requiring the number component of the
product ID will make the program more user-friendly.

Let's implement the creation of the product ID String from just an
inputted number.