Add proper authorship to code
The auto-generated authorship is inaccurate, as some members' code
have been overwritten in bug fixes. Thus, they lose authorship of
that portion of the code.

Explicitly stating the authorship in the code can help to clarify
doubts on who did which parts of the code.

Let's explicitly indicate authorship on the code.