Modify code after integrating upstream
Some sections of the code were not merged correctly after pulling
from the upstream master.

Properly integrating the upstream master will allow the code to
work seamlessly with the additional functionality that has been
added to the upstream master.

Let's check the code for merging issues, and correct them if
necessary.