Check JUnit tests
Since a significant portion of the code has been altered, the JUnit
tests may not as accurately test the methods in the code as before.

Checking and modifying the JUnit tests can help to maintain the
robustness of the JUnit tests in checking that the methods function
as intended.

Let's check the JUnit tests, and get them up-to-date if necessary.