Modify code after merging
After merging from the upstream master, some portions of the code
needs to be altered to follow the standards enforced in other
portions of the code.

Enforcing these code standards will allow the code to be more
consistent, which can improve readability, maintainability and
usability.

Let's check the edits, and make some changes to them if necessary.