Organise AboutUs.md and create PPP
The AboutUs.md needs some adjustments. We also each need a Personal
Portfolio Page (PPP) for this project.

Checking the AboutUs.md helps to make sure that it is acceptable to
be viewed by visitors on the Internet. Adding a PPP helps to break
down our individual contributions to the project.

Let's check the AboutUs.md file, and create a PPP for myself.