Modify code after merging
The code is not consistent with code quality standards. Moreover,
the explicit authorship is not accurate as expected.

Maintaining a consistent code quality can help to increase
readability and maintainability of the code.

Let's check the code for code quality issues, and modify authorship
for better recording of who did what.