Modify code after merging
The code is not consistent with code quality standards. Moreover,
there are some bugs in the newly-merged code.

Maintaining a consistent code quality can help to increase
readability and maintainability of the code. Fixing the bugs now
reduces the amount of regression testing we need to do, compared to
if fixing at a later time.

Let's check the code for code quality issues, and fix the bugs in
the newly-merged code.