Fix some bugs and update build.gradle
There are some bugs involving the quantitySold values. The first bug
is when the user inputs cumulative sales, the list is updated with
the most recent quantitySold value rather than the total amount of
quantitySold. The second bug is that quantitySold does not carry
over when saving the list data to the text file. Moreover, the
build.gradle needs to be updated to v2.0.

Fixing these bugs will ensure a better user experience, as the
behaviour of Busyness Manager will become more predictable.

Let's
* Fix these bugs.
* Edit the build.gradle file to v2.0.