Fix Bug #171
There is a bug reported, where the user does not recognise a user
story as the "print" command. This may be because the use of the
word "print" may not be clear enough to indicate its purpose of
printing out the product list.

Changing the "print" command to "list" may help in this case, as it
is now more obvious that this command's purpose is to print the
product list.

Let's change "print" to "list", and update all relevant
documentation.