Fix Bug #172
There is a bug reported, where the sales of a product can be
cleared, even when the quantity sold is 0. This is because the
method for clearing sales considers the case of quantitySold = 0 as
equivalent to if quantitySold > 0.

Outputting an error message when quantitySold = 0 will give users a
clearer understanding that the quantity sold for that product is 0,
thereby preventing them from getting confused as to why the sales
can still be cleared even when the quantity sold is zero.

Let's incorporate error handling for the case when the quantity
sold for a product is 0, and the user wants to clear the sales of
that product.