Fix Bug #174
There is a bug reported, where the user cannot add a product that
has a name with more than one word. This is due to our extraction
of attributes method that restricts the name to one word only.

Expanding the scope of the name to include more than one word will
help to make Busyness Manager more user-friendly. However, it is
not possible to do this while still using split(), which means that
we need to revamp the entire extraction of attributes method. This
will be time-consuming, and potentially buggy if we do not test it
well.

Let's continue with the one-word name approach, and just explain in
the documentation that the name of a product cannot be more than one
word, as well as why this is so.