Fix Bug #174
There is a bug reported, where the user mentioned that product IDs
are not reusable in the program. This is because the IDs functions
as keys in the inventory HashMap.

Making the product IDs reusable can help to improve functionality
effectivesness. However, this may cause issues after long-term
usage of the program.

Let's justify in the User Guide why the product IDs should not be
reused after deletion.