Troubleshoot Gradle check
The Gradle check is failing, as the expected outputs from the
automated text UI tests differ from the actual output.

Verifying that the automated text UI tests' expected outputs match
the actual outputs enables our program to be regression tested,
which checks that the functionality within the program works.

Let's edit the expected outputs of the automated text UI tests, and
check if the Gradle check passes.