Check code before release
Before releasing the next version, the code needs to be tidied up.

Checking the code will ensure that what we created is correct and
working. It will also enable us to find any last-minute bugs for
correction.

Let's look through the working and test code, and check for any
irregularities.